Reinstate dual facilities title approach.
Someone (Stu!) decided to remove the dual title approach (eg: _1 and _2) and replace with a single field. However they then continued to use the dual field approach in the template. As a result we had data being pulled through from the DB (which still existed) but no way to edit it on the live site.

This commit reinstates the dual field approach and removes the single field approach. A database backup will be taken prior to deploy.